5. A Project Requires An Owner
It's true that we can now create and persist projects to the database, but they aren't currently associated with any user. This isn't practical. To fix this, we'll write a test to confirm that the authenticated user is always assigned as the owner of any new project that is created during their session.

Creates authentication by:
1. php artisan make:auth
2. auth()->user()->projects()->create($attributes);